+++ scripts/macros/hdywtdt.js
