+++ scripts/types/Dice.d.ts
@@ -1,0 +1,6 @@
+interface DiceTrackParse {
+  name: string;
+  flavor: string;
+  isCritical: boolean;
+  isFumble: boolean;
+}
